Add "logFileRelativeUrl" to executeTestPlan
