+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@jsenv/core",
-  "version": "25.4.9",
+  "version": "25.5.0",
   "description": "Tool to develop, test and build js projects",
   "license": "MIT",
   "repository": {
@@ -101,6 +101,7 @@
     "rollup-plugin-node-globals": "1.4.0",
     "rollup-plugin-polyfill-node": "0.8.0",
     "source-map": "0.7.3",
+    "strip-ansi": "7.0.1",
     "systemjs": "6.11.0",
     "terser": "5.10.0",
     "v8-to-istanbul": "8.1.0",
