+++ test/test_plan/test_plan_fail_fast/b.spec.js
@@ -1,0 +1,1 @@
+throw new Error("toto")
